build: split images between raw and final
We need raw images into which we can move the build files, without
adjusted permissions. We need final images in which we can adjust
permissions. If we use the same image, then we won't be able to move
files to an incremental image because of file and directory permissions
having been adjusted to user 'ceph', 'cephadm', and others; users we
don't have on the host system, and for which the builder scripts do not
have permissions.

Signed-off-by: Joao Eduardo Luis <joao@suse.com>