cab: build base image, install requirements
Have the tool build the base image if missing, instead of forcing the
user to rely on an external tool for that purpose.

Also ensure that all known dependencies and recommended packages are
installed directly in this image, rather than having some of them
installed after each iteration -- thus saving time and raw space on the
fs.

Signed-off-by: Joao Eduardo Luis <joao@suse.com>